Move Sass task to separate file
